move angular material module
